address PR review comments: fix Open3.capture3 misuse, validate status param, and harden command execution

+++ app/services/make_fresh_downloadable.rb
@@ -76,7 +76,9 @@ class MakeFreshDownloadable < ApplicationService
           # TODO: figure out how not to go through epub
           epubname = make_epub_from_single_html(html, download_entity, author_string)
           mobiname = epubname[epubname.rindex('/') + 1..-6] + '.mobi'
-          system('kindlegen', epubname, '-c1', '-o', mobiname)
+          unless system('kindlegen', epubname, '-c1', '-o', mobiname)
+            raise "Kindlegen conversion failed for EPUB #{epubname}"
+          end
           mobiname = epubname[0..-6] + '.mobi'
           dl.stored_file.attach(io: File.open(mobiname), filename: filename)
           File.delete(epubname) # delete temporary generated EPUB
